requirements.txt
Flask>=2.0
python-dotenv>=1.0
oracledb>=1.0
bcrypt>=4.0
pytest>=7.0
